+++ tests/spec/libsass_closed_issues/mod.rs
@@ -11492,7 +11492,6 @@ fn issue_702() {
 
 // From "sass-spec/spec/libsass-closed-issues/issue_703.hrx"
 #[test]
-#[ignore] // wrong result
 fn issue_703() {
     assert_eq!(
         rsass(